fix analytics generate

+++ src/services/analytics/internal/proto/generate.sh
@@ -4,4 +4,7 @@ go install google.golang.org/protobuf/cmd/protoc-gen-go@latest
 protoc -I=./ --plugin=$(go env GOPATH)/bin/protoc-gen-go --plugin=../../../../common/proto/node_modules/.bin/protoc-gen-ts_proto --ts_proto_opt=esModuleInterop=true --go_out=../../ --ts_proto_out=../../../../packages/varsoapi/src/analytics base.proto requests.proto
 
 mkdir -p ../../../www/src/proto/src/analytics
-cp -R ../../../../packages/varsoapi/src/analytics ../../../www/src/proto/src/
+
+if [ -e "../../../../packages/varsoapi/src/analytics" ] ; then 
+  cp -R ../../../../packages/varsoapi/src/analytics ../../../www/src/proto/src/
+fi
